requirements.txt
mysql-connector-python
cx_Oracle
numpy
pyodbc